requirements.txt
numpy
pandas
seaborn
streamlit
pyYAML
scikit-learn
tensorflow
-e .